feat(plugins): move ALB related classes to keel-ec2-api
