+++ keel-clouddriver/src/main/kotlin/com/netflix/spinnaker/keel/clouddriver/model/ApplicationLoadBalancerModel.kt
@@ -1,8 +1,10 @@
 package com.netflix.spinnaker.keel.clouddriver.model
 
-import com.fasterxml.jackson.annotation.JsonAlias
 import com.fasterxml.jackson.annotation.JsonAnyGetter
 import com.netflix.spinnaker.keel.api.Moniker
+import com.netflix.spinnaker.keel.api.ec2.Action
+import com.netflix.spinnaker.keel.api.ec2.Rule
+import com.netflix.spinnaker.keel.api.ec2.TargetGroupAttributes
 
 data class ApplicationLoadBalancerModel(
   override val moniker: Moniker?,
@@ -31,34 +33,6 @@ data class ApplicationLoadBalancerModel(
     val certificateArn: String
   )
 
-  data class Action(
-    val type: String,
-    val order: Int,
-    val targetGroupName: String?,
-    val redirectConfig: RedirectConfig?
-  )
-
-  data class Rule(
-    val priority: String,
-    val conditions: List<Condition>?,
-    val actions: List<Action>,
-    val default: Boolean
-  )
-
-  data class Condition(
-    val field: String,
-    val values: List<String>
-  )
-
-  data class RedirectConfig(
-    val protocol: String,
-    val port: String?,
-    val host: String,
-    val path: String,
-    val query: String?,
-    val statusCode: String
-  )
-
   data class TargetGroup(
     val targetGroupName: String,
     val loadBalancerNames: List<String>,
@@ -81,24 +55,5 @@ data class ApplicationLoadBalancerModel(
 
   data class TargetGroupMatcher(
     val httpCode: String
-  )
-
-  data class TargetGroupAttributes(
-    @JsonAlias("stickiness.enabled")
-    val stickinessEnabled: Boolean = false,
-
-    @JsonAlias("deregistration_delay.timeout_seconds")
-    val deregistrationDelay: Int = 300,
-
-    @JsonAlias("stickiness.type")
-    val stickinessType: String = "lb_cookie",
-
-    @JsonAlias("stickiness.lb_cookie.duration_seconds")
-    val stickinessDuration: Int = 86400,
-
-    @JsonAlias("slow_start.duration_seconds")
-    val slowStartDurationSeconds: Int = 0,
-
-    @get:JsonAnyGetter val properties: Map<String, Any?> = emptyMap()
   )
 }
