+++ keel-ec2-api/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/Rule.kt
@@ -1,0 +1,8 @@
+package com.netflix.spinnaker.keel.api.ec2
+
+data class Rule(
+  val priority: String,
+  val conditions: List<Condition>?,
+  val actions: List<Action>,
+  val default: Boolean
+)
