+++ keel-ec2-api/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/old/ApplicationLoadBalancerV1Spec.kt
@@ -1,22 +1,25 @@
-package com.netflix.spinnaker.keel.api.ec2
+package com.netflix.spinnaker.keel.api.ec2.old
 
 import com.netflix.spinnaker.keel.api.Moniker
 import com.netflix.spinnaker.keel.api.SubnetAwareLocations
 import com.netflix.spinnaker.keel.api.UnhappyControl
+import com.netflix.spinnaker.keel.api.ec2.ApplicationLoadBalancerOverride
+import com.netflix.spinnaker.keel.api.ec2.ApplicationLoadBalancerSpec.Listener
+import com.netflix.spinnaker.keel.api.ec2.LoadBalancerDependencies
+import com.netflix.spinnaker.keel.api.ec2.LoadBalancerSpec
+import com.netflix.spinnaker.keel.api.ec2.LoadBalancerType
 import com.netflix.spinnaker.keel.api.ec2.LoadBalancerType.APPLICATION
-import com.netflix.spinnaker.keel.clouddriver.model.ApplicationLoadBalancerModel.Action
-import com.netflix.spinnaker.keel.clouddriver.model.ApplicationLoadBalancerModel.Rule
-import com.netflix.spinnaker.keel.clouddriver.model.ApplicationLoadBalancerModel.TargetGroupAttributes
+import com.netflix.spinnaker.keel.api.ec2.TargetGroupAttributes
 import java.time.Duration
 
-data class ApplicationLoadBalancerSpec(
+data class ApplicationLoadBalancerV1Spec(
   override val moniker: Moniker,
   override val locations: SubnetAwareLocations,
   override val internal: Boolean = true,
   override val dependencies: LoadBalancerDependencies = LoadBalancerDependencies(),
   override val idleTimeout: Duration = Duration.ofSeconds(60),
   val listeners: Set<Listener>,
-  val targetGroups: Set<TargetGroup>,
+  val targetGroups: Set<TargetGroupV1>,
   val overrides: Map<String, ApplicationLoadBalancerOverride> = emptyMap()
 ) : LoadBalancerSpec, UnhappyControl {
 
@@ -35,26 +38,18 @@ data class ApplicationLoadBalancerSpec(
 
   override val id: String = "${locations.account}:$moniker"
 
-  data class Listener(
-    val port: Int,
-    val protocol: String,
-    val certificateArn: String?,
-    val rules: Set<Rule> = emptySet(),
-    val defaultActions: Set<Action> = emptySet()
-  )
-
-  data class TargetGroup(
+  data class TargetGroupV1(
     val name: String,
     val targetType: String = "instance",
     val protocol: String = "HTTP",
     val port: Int,
     val healthCheckEnabled: Boolean = true,
-    val healthCheckTimeout: Duration = Duration.ofSeconds(5),
+    val healthCheckTimeoutSeconds: Duration = Duration.ofSeconds(5),
     val healthCheckPort: Int = 7001,
     val healthCheckProtocol: String = "HTTP",
     val healthCheckHttpCode: String = "200-299",
     val healthCheckPath: String = "/healthcheck",
-    val healthCheckInterval: Duration = Duration.ofSeconds(10),
+    val healthCheckIntervalSeconds: Duration = Duration.ofSeconds(10),
     val healthyThresholdCount: Int = 10,
     val unhealthyThresholdCount: Int = 2,
     val attributes: TargetGroupAttributes = TargetGroupAttributes()
@@ -66,9 +61,3 @@ data class ApplicationLoadBalancerSpec(
     }
   }
 }
-
-data class ApplicationLoadBalancerOverride(
-  val dependencies: LoadBalancerDependencies? = null,
-  val listeners: Set<ApplicationLoadBalancerSpec.Listener>? = null,
-  val targetGroups: Set<ApplicationLoadBalancerSpec.TargetGroup>? = null
-)
