+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/resolvers/ApplicationLoadBalancerDefaultsResolver.kt
@@ -1,9 +1,9 @@
 package com.netflix.spinnaker.keel.ec2.resolvers
 
 import com.netflix.spinnaker.keel.api.Resource
+import com.netflix.spinnaker.keel.api.ec2.Action
 import com.netflix.spinnaker.keel.api.ec2.ApplicationLoadBalancerSpec
 import com.netflix.spinnaker.keel.api.plugins.Resolver
-import com.netflix.spinnaker.keel.clouddriver.model.ApplicationLoadBalancerModel
 import com.netflix.spinnaker.keel.ec2.EC2_APPLICATION_LOAD_BALANCER_V1_1
 import org.springframework.stereotype.Component
 
@@ -17,7 +17,7 @@ class ApplicationLoadBalancerDefaultsResolver : Resolver<ApplicationLoadBalancer
         if (it.defaultActions.isEmpty()) {
           val defaultActions = if (it.defaultActions.isEmpty()) {
             setOf(
-              ApplicationLoadBalancerModel.Action(
+              Action(
                 type = "forward",
                 order = 1,
                 targetGroupName = resource.spec.targetGroups.first().name,
